Resolves query timeout issue with the call to purge deleted items from the manifest.

+++ mill-db-repo/src/main/java/org/duracloud/mill/manifest/ManifestStore.java
@@ -98,7 +98,7 @@ public interface ManifestStore {
      * @param expiration
      * @return Count of items deleted.
      */
-    public Long purgeDeletedItemsBefore(Date expiration);
+    public int purgeDeletedItemsBefore(Date expiration);
 
     /**
      *
